+++ crates/threshold-signature-server/src/helpers/app_state.rs
@@ -19,19 +19,19 @@ use crate::{
     launch::Configuration,
     signing_client::ListenerState,
 };
-use anyhow::anyhow;
 use entropy_kvdb::kv_manager::KvManager;
 use entropy_shared::X25519PublicKey;
 use serde::{Deserialize, Serialize};
 use sp_core::{crypto::AccountId32, sr25519, Pair};
 use std::{
     collections::HashMap,
-    sync::{Arc, PoisonError, RwLock},
+    sync::{Arc, RwLock},
 };
 use subxt::{
     backend::legacy::LegacyRpcMethods, tx::PairSigner, utils::AccountId32 as SubxtAccountId32,
     OnlineClient,
 };
+use thiserror::Error;
 use x25519_dalek::StaticSecret;
 
 /// Represents the state relating to the prerequisite checks
@@ -138,10 +138,9 @@ impl Cache {
     }
 
     /// Mark the node as able to make chain queries. This is called once during prerequisite checks
-    pub fn connected_to_chain_node(
-        &self,
-    ) -> Result<(), PoisonError<std::sync::RwLockWriteGuard<'_, TssState>>> {
-        let mut tss_state = self.tss_state.write()?;
+    pub fn connected_to_chain_node(&self) -> Result<(), AppStateError> {
+        let mut tss_state =
+            self.tss_state.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         if *tss_state == TssState::NoChainConnection {
             *tss_state = TssState::ReadOnlyChainConnection;
         }
@@ -149,61 +148,52 @@ impl Cache {
     }
 
     /// Mark the node as ready. This is called once when the prerequisite checks have passed.
-    pub fn make_ready(&self) -> Result<(), PoisonError<std::sync::RwLockWriteGuard<'_, TssState>>> {
-        let mut tss_state = self.tss_state.write()?;
+    pub fn make_ready(&self) -> Result<(), AppStateError> {
+        let mut tss_state =
+            self.tss_state.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         *tss_state = TssState::Ready;
         Ok(())
     }
 
     /// Write to request limit
-    pub fn write_to_request_limit(&self, key: String, value: u32) -> anyhow::Result<()> {
+    pub fn write_to_request_limit(&self, key: String, value: u32) -> Result<(), AppStateError> {
         self.clear_poisioned_request_limit();
-        let mut request_limit = self
-            .request_limit
-            .write()
-            .map_err(|_| anyhow!("Error getting write write_to_request_limit lock"))?;
+        let mut request_limit =
+            self.request_limit.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         request_limit.insert(key, value);
         Ok(())
     }
 
     /// Check if key exists in request limit
-    pub fn exists_in_request_limit(&self, key: &String) -> anyhow::Result<bool> {
+    pub fn exists_in_request_limit(&self, key: &String) -> Result<bool, AppStateError> {
         self.clear_poisioned_request_limit();
-        let request_limit = self
-            .request_limit
-            .read()
-            .map_err(|_| anyhow!("Error getting read exists_in_request_limit lock"))?;
+        let request_limit =
+            self.request_limit.read().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         Ok(request_limit.contains_key(key))
     }
 
     /// Remove key from request limt
-    pub fn remove_from_request_limit(&self, key: &String) -> anyhow::Result<()> {
+    pub fn remove_from_request_limit(&self, key: &String) -> Result<(), AppStateError> {
         self.clear_poisioned_request_limit();
-        let mut request_limit = self
-            .request_limit
-            .write()
-            .map_err(|_| anyhow!("Error getting write remove_from_request_limit lock"))?;
+        let mut request_limit =
+            self.request_limit.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         request_limit.remove(key);
         Ok(())
     }
 
     /// Reads from request_limit will error if no value, call exists_in_request_limit to check
-    pub fn read_from_request_limit(&self, key: &String) -> anyhow::Result<Option<u32>> {
+    pub fn read_from_request_limit(&self, key: &String) -> Result<Option<u32>, AppStateError> {
         self.clear_poisioned_request_limit();
-        let request_limit = self
-            .request_limit
-            .read()
-            .map_err(|_| anyhow!("Error getting read read_from_request_limit lock"))?;
+        let request_limit =
+            self.request_limit.read().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         Ok(request_limit.get(key).cloned())
     }
 
     /// Clears the request_limit mapping
-    pub fn clear_request_limit(&self) -> anyhow::Result<()> {
+    pub fn clear_request_limit(&self) -> Result<(), AppStateError> {
         self.clear_poisioned_request_limit();
-        let mut request_limit = self
-            .request_limit
-            .write()
-            .map_err(|_| anyhow!("Error getting read read_from_request_limit lock"))?;
+        let mut request_limit =
+            self.request_limit.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         request_limit.clear();
         Ok(())
     }
@@ -216,23 +206,25 @@ impl Cache {
     }
 
     /// Write the given block number to the `block_number` cache.
-    pub fn write_to_block_numbers(&self, key: BlockNumberFields, value: u32) -> anyhow::Result<()> {
+    pub fn write_to_block_numbers(
+        &self,
+        key: BlockNumberFields,
+        value: u32,
+    ) -> Result<(), AppStateError> {
         let block_number_target = self.get_block_number_target(&key);
         self.clear_poisioned_block_numbers(&block_number_target);
-        let mut block_number = block_number_target
-            .write()
-            .map_err(|_| anyhow!("Error getting write from write_to_block_numbers lock"))?;
+        let mut block_number =
+            block_number_target.write().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         *block_number = value;
         Ok(())
     }
 
     /// Returns the number of requests handled so far at the given block number.
-    pub fn read_from_block_numbers(&self, key: &BlockNumberFields) -> anyhow::Result<u32> {
+    pub fn read_from_block_numbers(&self, key: &BlockNumberFields) -> Result<u32, AppStateError> {
         let block_number_target = self.get_block_number_target(key);
         self.clear_poisioned_block_numbers(&block_number_target);
-        let block_number = block_number_target
-            .read()
-            .map_err(|_| anyhow!("Error getting read from read_to_block_numbers lock"))?;
+        let block_number =
+            block_number_target.read().map_err(|e| AppStateError::PosionError(e.to_string()))?;
         Ok(*block_number)
     }
 
@@ -257,9 +249,9 @@ impl Cache {
     pub fn unsubscribed_peers(
         &self,
         session_id: &entropy_protocol::SessionId,
-    ) -> Result<Vec<subxt::utils::AccountId32>, crate::signing_client::ProtocolErr> {
+    ) -> Result<Vec<subxt::utils::AccountId32>, AppStateError> {
         self.listener_state.unsubscribed_peers(session_id).map_err(|_| {
-            crate::signing_client::ProtocolErr::SessionError(format!(
+            AppStateError::SessionError(format!(
                 "Unable to get unsubscribed peers for `SessionId` {:?}",
                 session_id,
             ))
@@ -295,7 +287,7 @@ impl AppState {
     /// Convenience function to get chain api and rpc
     pub async fn get_api_rpc(
         &self,
-    ) -> Result<(OnlineClient<EntropyConfig>, LegacyRpcMethods<EntropyConfig>), subxt::Error> {
+    ) -> Result<(OnlineClient<EntropyConfig>, LegacyRpcMethods<EntropyConfig>), AppStateError> {
         Ok((
             get_api(&self.configuration.endpoint).await?,
             get_rpc(&self.configuration.endpoint).await?,
@@ -321,4 +313,15 @@ impl AppState {
     pub fn x25519_public_key(&self) -> [u8; 32] {
         x25519_dalek::PublicKey::from(&self.x25519_secret).to_bytes()
     }
+}
+
+/// Errors related to app state.
+#[derive(Error, Debug)]
+pub enum AppStateError {
+    #[error("Posion Mutex error: {0}")]
+    PosionError(String),
+    #[error("Session Error: {0}")]
+    SessionError(String),
+    #[error("Subxt: {0}")]
+    Subxt(#[from] subxt::Error),
 }