+++ crates/threshold-signature-server/src/validator/tests.rs
@@ -22,6 +22,7 @@ use crate::{
         },
     },
     validator::api::{is_signer_or_delete_parent_key, prune_old_holders, validate_new_reshare},
+    EntropyConfig,
 };
 use entropy_client::{self as test_client};
 use entropy_client::{
@@ -31,7 +32,9 @@ use entropy_client::{
         entropy::runtime_types::entropy_runtime::RuntimeCall,
         entropy::runtime_types::frame_system::pallet::Call as SystemsCall,
         entropy::runtime_types::pallet_registry::pallet::ProgramInstance,
-        entropy::runtime_types::pallet_staking_extension::pallet::{NextSignerInfo, ReshareInfo},
+        entropy::runtime_types::pallet_staking_extension::pallet::{
+            NextSignerInfo, ReshareInfo, ServerInfo,
+        },
         get_api, get_rpc,
     },
     substrate::query_chain,
@@ -52,12 +55,9 @@ use serial_test::serial;
 use sp_core::Pair;
 use sp_keyring::AccountKeyring;
 use std::collections::HashSet;
-use subxt::utils::AccountId32;
+use subxt::{backend::legacy::LegacyRpcMethods, utils::AccountId32, OnlineClient};
 use synedrion::k256::ecdsa::VerifyingKey;
 
-// FIXME (#1273): This fails intermittently and needs to be addressed. For now we ignore it since
-// it's producing false negatives on our CI runs.
-#[ignore]
 #[tokio::test]
 #[serial]
 async fn test_reshare_basic() {
@@ -67,7 +67,6 @@ async fn test_reshare_basic() {
     let (_validator_ips, _validator_ids) =
         spawn_testing_validators(crate::helpers::tests::ChainSpecType::IntegrationJumpStarted)
             .await;
-
     let force_authoring = true;
     let context =
         test_node_process_testing_state(ChainSpecType::IntegrationJumpStarted, force_authoring)
@@ -83,13 +82,10 @@ async fn test_reshare_basic() {
     let signer_stash_accounts = query_chain(&api, &rpc, signer_query, None).await.unwrap().unwrap();
     let old_signer_ids: HashSet<[u8; 32]> =
         HashSet::from_iter(signer_stash_accounts.clone().into_iter().map(|id| id.0));
-    let mut signers = Vec::new();
+    let signers = get_current_signers(&api, &rpc).await;
     let mut next_signers = vec![];
     for signer in signer_stash_accounts.iter() {
         next_signers.push(signer);
-        let query = entropy::storage().staking_extension().threshold_servers(signer);
-        let server_info = query_chain(&api, &rpc, query, None).await.unwrap().unwrap();
-        signers.push(server_info);
     }
 
     for signer in signers.iter() {
@@ -119,7 +115,7 @@ async fn test_reshare_basic() {
     });
     call_set_storage(&api, &rpc, call).await;
 
-    let key_share_before = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3002).await;
+    let key_shares_before = get_all_keys(signers).await;
 
     let mut i = 0;
     // Wait up to 2min for reshare to complete: check once every second if we have a new set of signers.
@@ -142,9 +138,11 @@ async fn test_reshare_basic() {
 
     // wait for roatate keyshare
     tokio::time::sleep(std::time::Duration::from_secs(10)).await;
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_shares_after = get_all_keys(signers).await;
 
-    let key_share_after = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3002).await;
-    assert_ne!(key_share_before, key_share_after);
+    assert_ne!(key_shares_before, key_shares_after);
+
     // At this point the signing set has changed on-chain, but the keyshares haven't been rotated
     // but by the time we have stored a program and registered, the rotation should have happened
 
@@ -210,7 +208,8 @@ async fn test_reshare_basic() {
         assert!(!key_share.is_empty());
     }
     let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
-    let key_share_before_2 = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3003).await;
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_share_before_2 = get_all_keys(signers).await;
 
     next_signers.remove(0);
     let binding = alice_stash.to_account_id().into();
@@ -236,7 +235,9 @@ async fn test_reshare_basic() {
     // wait for roatate keyshare
     tokio::time::sleep(std::time::Duration::from_secs(60)).await;
 
-    let key_share_after_2 = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3003).await;
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_share_after_2 = get_all_keys(signers).await;
+
     assert_ne!(key_share_before_2, key_share_after_2);
 
     clean_tests();
@@ -259,14 +260,14 @@ async fn test_reshare_e2e() {
             .await;
     let api = get_api(&context[0].ws_url).await.unwrap();
     let rpc = get_rpc(&context[0].ws_url).await.unwrap();
-    let client = reqwest::Client::new();
 
     // Get current signers
     let signer_query = entropy::storage().staking_extension().signers();
     let signer_stash_accounts = query_chain(&api, &rpc, signer_query, None).await.unwrap().unwrap();
     let old_signer_ids: HashSet<[u8; 32]> =
         HashSet::from_iter(signer_stash_accounts.clone().into_iter().map(|id| id.0));
-    let key_share_before = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3002).await;
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_share_before = get_all_keys(signers).await;
 
     let mut i = 0;
     // Wait up to 2min for reshare to complete: check once every second if we have a new set of signers.
@@ -289,10 +290,13 @@ async fn test_reshare_e2e() {
     // wait for rotate keyshare
     tokio::time::sleep(std::time::Duration::from_secs(10)).await;
 
-    let key_share_after = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3002).await;
-    assert_ne!(key_share_before, key_share_after);
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_shares_after = get_all_keys(signers).await;
 
-    let key_share_before_2 = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3003).await;
+    assert_ne!(key_share_before, key_shares_after);
+
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_share_before_2 = get_all_keys(signers).await;
 
     let _ = loop {
         let new_signer_ids: HashSet<[u8; 32]> = {
@@ -314,7 +318,9 @@ async fn test_reshare_e2e() {
     // wait for rotate keyshare 2
     tokio::time::sleep(std::time::Duration::from_secs(10)).await;
 
-    let key_share_after_2 = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), 3003).await;
+    let signers = get_current_signers(&api, &rpc).await;
+    let key_share_after_2 = get_all_keys(signers).await;
+
     assert_ne!(key_share_before_2, key_share_after_2);
 }
 
@@ -437,6 +443,7 @@ async fn test_empty_next_signer() {
 }
 
 #[tokio::test]
+#[serial]
 #[should_panic = "Account does not exist, add balance"]
 async fn test_check_balance_for_fees() {
     initialize_test_logger().await;
@@ -483,4 +490,34 @@ async fn test_deletes_key() {
     let has_key = kv.kv().exists(&hex::encode(NETWORK_PARENT_KEY)).await.unwrap();
     assert!(!has_key);
     clean_tests();
+}
+
+/// Get all the network keys from the server info provided
+pub async fn get_all_keys(servers_info: Vec<ServerInfo<AccountId32>>) -> HashSet<Vec<u8>> {
+    let client = reqwest::Client::new();
+    let mut key_shares = vec![];
+    for server_info in servers_info {
+        let port = get_port(&server_info);
+        let result = unsafe_get(&client, hex::encode(NETWORK_PARENT_KEY), port).await;
+        if !result.is_empty() {
+            key_shares.push(result);
+        }
+    }
+    HashSet::from_iter(key_shares)
+}
+
+/// Gets the current signers server info
+pub async fn get_current_signers(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+) -> Vec<ServerInfo<AccountId32>> {
+    let signer_query = entropy::storage().staking_extension().signers();
+    let signer_stash_accounts = query_chain(&api, &rpc, signer_query, None).await.unwrap().unwrap();
+    let mut signers = Vec::new();
+    for signer in signer_stash_accounts.iter() {
+        let query = entropy::storage().staking_extension().threshold_servers(signer);
+        let server_info = query_chain(&api, &rpc, query, None).await.unwrap().unwrap();
+        signers.push(server_info);
+    }
+    signers
 }